@@ -18,27 +18,27 @@
     "testSR25519": "jest --detectOpenHandles --runInBand tests/sr25519.test.ts"
   },
   "dependencies": {
-    "@zondax/ledger-polkadot": "^0.13.6",
-    "@zondax/zemu": "^0.12.3"
+    "@zondax/ledger-polkadot": "^0.14.0",
+    "@zondax/zemu": "^0.14.5"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.22",
+    "@types/jest": "^26.0.23",
     "@types/ledgerhq__hw-transport": "^4.21.3",
-    "@typescript-eslint/eslint-plugin": "^4.20.0",
-    "@typescript-eslint/parser": "^4.20.0",
+    "@typescript-eslint/eslint-plugin": "^4.24.0",
+    "@typescript-eslint/parser": "^4.24.0",
     "blakejs": "^1.1.0",
     "crypto-js": "4.0.0",
     "ed25519-supercop": "^2.0.1",
-    "eslint": "^7.24.0",
-    "eslint-config-prettier": "^8.2.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.3.5",
-    "eslint-plugin-prettier": "^3.4.0",
+    "eslint": "^7.26.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.23.2",
+    "eslint-plugin-jest": "^24.3.6",
+    "eslint-plugin-prettier": "^3.3.1",
     "jest": "26.6.3",
     "jest-serial-runner": "^1.1.0",
     "jssha": "^3.1.1",
-    "prettier": "^2.0.4",
-    "ts-jest": "^26.5.4",
+    "prettier": "^2.3.0",
+    "ts-jest": "^26.5.6",
     "typescript": "^4.2.3"
   }
 }
