@@ -42,6 +42,7 @@ void h_secret_click();
 ux_state_t ux;
 
 void os_exit(uint32_t id) {
+    (void)id;
     os_sched_exit(0);
 }
 
@@ -58,7 +59,6 @@ const ux_menu_entry_t menu_main[] = {
 #endif
      0, &C_icon_app, "Developed by:", "Zondax.ch", 33, 12},
 
-    {NULL, NULL, 0, &C_icon_app, "Developed by:", "Zondax.ch", 33, 12},
     {NULL, NULL, 0, &C_icon_app, "License: ", "Apache 2.0", 33, 12},
     {NULL, os_exit, 0, &C_icon_dashboard, "Quit", NULL, 50, 29},
     UX_MENU_END
@@ -86,6 +86,7 @@ static const bagl_element_t view_error[] = {
 };
 
 static unsigned int view_error_button(unsigned int button_mask, unsigned int button_mask_counter) {
+    UNUSED(button_mask_counter);
     switch (button_mask) {
         case BUTTON_EVT_RELEASED | BUTTON_LEFT | BUTTON_RIGHT:
         case BUTTON_EVT_RELEASED | BUTTON_LEFT:
@@ -98,6 +99,7 @@ static unsigned int view_error_button(unsigned int button_mask, unsigned int but
 }
 
 static unsigned int view_message_button(unsigned int button_mask, unsigned int button_mask_counter) {
+    UNUSED(button_mask_counter);
     switch (button_mask) {
         case BUTTON_EVT_RELEASED | BUTTON_LEFT | BUTTON_RIGHT:
         case BUTTON_EVT_RELEASED | BUTTON_LEFT:
@@ -108,6 +110,7 @@ static unsigned int view_message_button(unsigned int button_mask, unsigned int b
 }
 
 static unsigned int view_review_button(unsigned int button_mask, unsigned int button_mask_counter) {
+    UNUSED(button_mask_counter);
     switch (button_mask) {
         case BUTTON_EVT_RELEASED | BUTTON_LEFT | BUTTON_RIGHT:
             h_review_button_both();
