@@ -18,7 +18,7 @@
 #include <string.h>
 #include "hexutils.h"
 
-uint8_t hex2dec(char c, uint8_t *out) {
+uint8_t hex2dec(char c, char *out) {
     c = (char) tolower((int) c);
 
     if (!isxdigit((int) c)) {
@@ -44,7 +44,7 @@ size_t parseHexString(uint8_t *out, uint16_t outLen, const char *input) {
     }
 
     for (size_t i = 0; i < len; i += 2) {
-        uint8_t tmp1, tmp2;
+        char tmp1, tmp2;
         if (hex2dec(input[i], &tmp1))
             return 0;
         if (hex2dec(input[i + 1], &tmp2))
